fix homework 1
