+++ src/main/java/company/vk/polis/ads/week1/Task1.java
@@ -1,3 +1,5 @@
+package company.vk.polis.ads.week1;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -5,8 +7,8 @@ import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.util.*;
 
-public final class Main {
-    private Main() {
+public final class Task1 {
+    private Task1() {
         // Should not be instantiated
     }
