+++ src/main/java/company/vk/polis/ads/week1/Task6.java
@@ -1,3 +1,4 @@
+package company.vk.polis.ads.week1;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,8 +7,8 @@ import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.util.*;
 
-public final class Main {
-    private Main() {
+public final class Task6 {
+    private Task6() {
         // Should not be instantiated
     }
